% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cs_demog}
\alias{cs_demog}
\title{Cross-sectional Demographic Data}
\format{A \code{data.frame} with 6 columns, which are:
\describe{
\item{id}{patient identifier}
\item{age}{age of patient}
\item{sex}{sex of the patient}
\item{ms_type}{type of MS diagnosis}
\item{edss}{Expanded Disability Status Scale (EDSS)}
\item{criteria}{criteria for diagnosis of MS}
}}
\usage{
cs_demog
}
\description{
Cross-sectional Demographic Data
}
\keyword{datasets}
