% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cs_raw}
\alias{cs_raw}
\title{Raw Cross-sectional Data}
\format{A \code{data.frame} with 3 columns, which are:
\describe{
\item{file}{filename, necessary for downloading data, along with the URL}
\item{mod}{imaging modality}
\item{id}{patient identifier}
}}
\usage{
cs_raw
}
\description{
Raw Cross-sectional Data
}
\keyword{datasets}
