% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_ms_data.R
\name{download_ms_patient}
\alias{download_ms_patient}
\alias{ms_patient_urls}
\title{Download MS Imaging Data}
\usage{
download_ms_patient(id, cohort = c("cross_sectional", "longitudinal"),
  data = c("raw", "coregistered"), outdir = tempfile(), overwrite = FALSE)

ms_patient_urls(id, cohort = c("cross_sectional", "longitudinal"),
  data = c("raw", "coregistered"))
}
\arguments{
\item{id}{ID to download (either 01 or patient01 or 1)}

\item{cohort}{Cohort to be used - cross-sectional or longitudinal}

\item{data}{data to be used - raw or coregistered/processed}

\item{outdir}{output directory}

\item{overwrite}{Should the files be overwritten if already existing?}
}
\value{
A \code{data.frame} of imaging modalities and masks
}
\description{
Download MS Imaging Data
}
\examples{
\dontrun{
res = download_ms_patient(id = "patient01")
res = download_ms_patient(
id = "patient01",
data = "coregistered",
cohort = "longitudinal")
}
}
