% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{long_coreg}
\alias{long_coreg}
\title{Coregistered Longitudinal Data}
\format{A \code{data.frame} with 5 columns, which are:
\describe{
\item{file}{filename, necessary for downloading data, along with the URL}
\item{mod}{imaging modality}
\item{id}{patient identifier}
\item{Brain_Mask}{filename of the brain mask}
\item{Gold_Standard}{filename of the gold standard lesion change}
}}
\usage{
long_coreg
}
\description{
Coregistered Longitudinal Data
}
\keyword{datasets}
