% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{long_demog}
\alias{long_demog}
\title{Longitudinal Demographic Data}
\format{A \code{data.frame} with 6 columns, which are:
\describe{
\item{patient_id}{patient identifier}
\item{age}{age of patient, at first visit}
\item{sex}{sex of the patient}
\item{ms_type}{type of MS diagnosis}
\item{days_between_studies}{number of days between study 1 and study 2}
}}
\usage{
long_demog
}
\description{
Longitudinal Demographic Data
}
\keyword{datasets}
