% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{long_raw}
\alias{long_raw}
\title{Raw Longitudinal Data}
\format{A \code{data.frame} with 4 columns, which are:
\describe{
\item{file}{filename, necessary for downloading data, along with the URL}
\item{mod}{imaging modality}
\item{id}{patient identifier}
\item{Gold_Standard}{filename of the gold standard lesion change}
}}
\usage{
long_raw
}
\description{
Raw Longitudinal Data
}
\keyword{datasets}
