% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms_data.R
\name{ms_data}
\alias{ms_data}
\title{MS Imaging Data in a \code{data.frame}}
\usage{
ms_data(cohort = c("cross_sectional", "longitudinal"), data = c("raw",
  "coregistered"))
}
\arguments{
\item{cohort}{Cohort to be used - cross-sectional or longitudinal}

\item{data}{data to be used - raw or coregistered/processed}
}
\value{
A \code{data.frame} of imaging modalities and masks
}
\description{
MS Imaging Data in a \code{data.frame}
}
\examples{
ms_data()
}
