% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms_demog.R
\name{ms_demog}
\alias{ms_demog}
\title{MS Demographic Data in a \code{data.frame}}
\usage{
ms_demog(cohort = c("cross_sectional", "longitudinal"))
}
\arguments{
\item{cohort}{Cohort to be used - cross-sectional or longitudinal}
}
\value{
A \code{data.frame} of patient demographics
}
\description{
MS Demographic Data in a \code{data.frame}
}
\examples{
ms_demog()
df = ms_demog(cohort = "longitudinal")
}
