% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms_ids.R
\name{ms_ids}
\alias{ms_ids}
\title{MS Data Ids}
\usage{
ms_ids(cohort = c("cross_sectional", "longitudinal"))
}
\arguments{
\item{cohort}{cohort (cross-sectional or longitudinal) of data}
}
\value{
A vector of IDs
}
\description{
MS Data Ids
}
\examples{
ms_ids()
ms_ids("longitudinal")
}
